requirements.txt
pandas
numpy
yfinance
scikit-learn
PyYAML
pytest
